+++ src/QuantumClifford.jl
@@ -95,7 +95,7 @@ export
     # petrajectories
     petrajectories, applybranches,
     # nonclifford
-    GeneralizedStabilizer, UnitaryPauliChannel, PauliChannel, pcT, pcPhase,
+    GeneralizedStabilizer, UnitaryPauliChannel, PauliChannel, pcT, pcPhase, pcRx,
     # makie plotting -- defined only when extension is loaded
     stabilizerplot, stabilizerplot_axis,
     # sum types
@@ -570,6 +570,13 @@ column permutation in the preparation of a `MixedDestabilizer` so that qubits ar
 The boolean keyword arguments `undoperm` and `reportperm` can be used to control this behavior
 and to report the permutations explicitly.
 
+Occasionally one might want specific destabilizer operators for a given tableau, **without** canonicalizing the tableau.
+E.g. if a Pauli frame correction is to be applied after purification or teleportation,
+the correction has to be done by applying the destabilizer corresponding to a specific stabilizer operator of a given tableau.
+Directly using `MixedDestabilizer` will first canonicalize the tableau, which would change the rows of the tableau,
+leading to a consistent set of destabilizer operators, but not the specific ones corresponding to the original pre-canonicalization tableau.
+The `backtrack=true` keyword argument can be undo canonicalization and restore the original rows, but this time also with the corresponding destabilizers.
+
 See also: [`stabilizerview`](@ref), [`destabilizerview`](@ref), [`logicalxview`](@ref), [`logicalzview`](@ref)
 """
 mutable struct MixedDestabilizer{T<:Tableau} <: AbstractStabilizer
@@ -578,9 +585,13 @@ mutable struct MixedDestabilizer{T<:Tableau} <: AbstractStabilizer
 end
 
 # Added a lot of type assertions to help Julia infer types
-function MixedDestabilizer(stab::Stabilizer{T}; undoperm=true, reportperm=false) where {T}
+function MixedDestabilizer(stab::Stabilizer{T}; undoperm=true, reportperm=false, backtrack=false) where {T}
     rows,n = size(stab)
-    stab, r, s, permx, permz = canonicalize_gott!(copy(stab))
+    if backtrack
+        stab, r, s, permx, permz, canonops = canonicalize_gott!(copy(stab), backtrack=true)
+    else
+        stab, r, s, permx, permz = canonicalize_gott!(copy(stab))
+    end
     t = zero(T, n*2, n)
     vstab = @view tab(stab)[1:r+s] # this view is necessary for cases of tableaux with redundant rows
     t[n+1:n+r+s] = vstab # The Stabilizer part of the tableau
@@ -615,12 +626,17 @@ function MixedDestabilizer(stab::Stabilizer{T}; undoperm=true, reportperm=false)
     end
     if undoperm
         t = t[:,invperm(permx[permz])]
-        return MixedDestabilizer(t, r+s)
+    end
+    returnstate = MixedDestabilizer(t, r+s)
+    if backtrack
+        for canonop in reverse(canonops)
+            canonop(returnstate)
+        end
     end
     if reportperm
-        return (MixedDestabilizer(t, r+s)::MixedDestabilizer{T}, r, permx, permz)
+        return returnstate, r, permx, permz
     else
-        return MixedDestabilizer(t, r+s)::MixedDestabilizer{T}
+        return returnstate
     end
 end